requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
plotly==5.6.0
PyYAML==5.3.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.63.0
kaleido==0.2.1